requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
fastapi
uvicorn
streamlit
pillow
python-multipart
tqdm
matplotlib
seaborn
requests
